+++ app/editor/src/hooks/api-editor/interfaces/IAdminActionModel.ts
@@ -1,0 +1,3 @@
+import { ISortableModel } from '.';
+
+export interface IAdminActionModel extends ISortableModel<number> {}
