+++ app/editor/src/hooks/api-editor/interfaces/index.ts
@@ -1,5 +1,6 @@
 export * from './IActionFilter';
 export * from './IActionModel';
+export * from './IAdminActionModel';
 export * from './IAuditColumnsModel';
 export * from './ICacheModel';
 export * from './ICategoryModel';
